backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy>=1.26.0
tensorflow>=2.15.0
scikit-learn>=1.3.2
pandas>=2.1.4
pymongo>=4.6.0
python-dotenv>=1.0.0
bcrypt>=4.1.2
pyjwt>=2.8.0
